Organize: move NPTMove to timemachine.md.barostat.moves
