+++ tests/test_mtm.py
@@ -11,7 +11,8 @@ from timemachine import testsystems
 from timemachine.constants import BOLTZ
 from timemachine.ff import Forcefield
 from timemachine.md import enhanced
-from timemachine.md.moves import NPTMove, OptimizedMTMMove, ReferenceMTMMove
+from timemachine.md.barostat.moves import NPTMove
+from timemachine.md.moves import OptimizedMTMMove, ReferenceMTMMove
 from timemachine.md.states import CoordsVelBox
 from timemachine.potentials import NonbondedInteractionGroup, nonbonded
 
@@ -136,7 +137,7 @@ def test_optimized_MTM():
     bps = [ubp.bind(params) for ubp, params in zip(ubps, params)]
 
     # we should initialize new instances of this
-    npt_mover = NPTMove(bps, masses, temperature, pressure, n_steps=1000, seed=seed)
+    npt_mover = NPTMove(bps, masses, temperature, pressure=pressure, n_steps=1000, seed=seed)
 
     K = 100
     # note that these seeds aren't actually used, since we feed in explicit keys to acceptance_probability
