+++ timemachine/md/moves.py
@@ -13,9 +13,8 @@ from scipy.special import logsumexp
 
 from timemachine import lib
 from timemachine.lib import custom_ops
-from timemachine.md.barostat.utils import get_bond_list, get_group_indices
 from timemachine.md.states import CoordsVelBox
-from timemachine.potentials import BoundPotential, HarmonicBond
+from timemachine.potentials import BoundPotential
 
 _State = TypeVar("_State")
 
@@ -113,43 +112,6 @@ class NVTMove(Move[CoordsVelBox]):
         after_steps = CoordsVelBox(x_t, v_t, box)
 
         return after_steps
-
-
-class NPTMove(NVTMove):
-    """
-    Functionally, NPT is implemented as NVTMove plus a MC Barostat.
-    So inherit from NVTMove here.
-    """
-
-    def __init__(
-        self,
-        bps: List[BoundPotential],
-        masses: NDArray,
-        temperature: float,
-        pressure: float,
-        n_steps: int,
-        seed: int,
-        dt: float = 1.5e-3,
-        friction: float = 1.0,
-        barostat_interval: int = 5,
-    ):
-        super().__init__(bps, masses, temperature, n_steps, seed, dt=dt, friction=friction)
-
-        assert isinstance(bps[0].potential, HarmonicBond), "First potential must be of type HarmonicBond"
-
-        bond_list = get_bond_list(bps[0].potential)
-        group_idxs = get_group_indices(bond_list, len(masses))
-
-        barostat = lib.MonteCarloBarostat(len(masses), pressure, temperature, group_idxs, barostat_interval, seed + 1)
-        barostat_impl = barostat.impl(self.bound_impls)
-        self.barostat_impl = barostat_impl
-
-    def move(self, x: CoordsVelBox) -> CoordsVelBox:
-        # note: context creation overhead here is actually very small!
-        ctxt = custom_ops.Context(
-            x.coords, x.velocities, x.box, self.integrator_impl, self.bound_impls, self.barostat_impl
-        )
-        return self._steps(ctxt)
 
 
 class DeterministicMTMMove(Move):
